Download stvResults
