+++ src/components/activeVotation/VotationResult.tsx
@@ -6,15 +6,17 @@ import {
   GetVotationResultsQuery,
   VotationType,
 } from '../../__generated__/graphql-types';
-import { Center, VStack, Text, Divider, Button } from '@chakra-ui/react';
+import { Center, VStack, Text, Divider, Button, HStack } from '@chakra-ui/react';
 import Hammer from '../../static/hammer.svg';
 import { ArrowBackIcon } from '@chakra-ui/icons';
 import ResultsTable from './results_table/ResultsTable';
 import Loading from '../common/Loading';
 import AlternativesString from '../common/AlternativesString';
 import StvResultTable from './results_table/StvResultTable';
+import DownloadResultButton from '../DownloadResultButton';
 
 export interface VotationResultProps {
+  isAdmin: boolean;
   winners: AlternativeType[] | AlternativeResult[] | null;
   result: GetVotationResultsQuery | null | undefined;
   votationId: string;
@@ -67,9 +69,14 @@ const VotationResult: React.FC<VotationResultProps> = ({
           <ResultsTable result={result} votationId={votationId} />
         ))}
       <Divider m="3em 0" />
-      <Button borderRadius={'16em'} onClick={backToVotationList} leftIcon={<ArrowBackIcon />}>
-        Gå tilbake til liste over voteringer
-      </Button>
+      <HStack w="100%" justifyContent="space-around">
+        {(result || stvResult) && (
+          <DownloadResultButton isStv={type === VotationType.Stv} result={result} stvResult={stvResult} />
+        )}
+        <Button borderRadius={'16em'} onClick={backToVotationList} leftIcon={<ArrowBackIcon />}>
+          Gå tilbake til liste over voteringer
+        </Button>
+      </HStack>
     </VStack>
   );
 };
